+++ djangosige/apps/financeiro/views/plano.py
@@ -1,18 +1,18 @@
 # -*- coding: utf-8 -*-
 
 from django.core.urlresolvers import reverse_lazy
-from django.views.generic.edit import CreateView, UpdateView
-from django.views.generic.base import TemplateView
-from django.contrib import messages
 from django.shortcuts import redirect
+
+from djangosige.apps.base.custom_views import CustomCreateView, CustomUpdateView, CustomTemplateView
 
 from djangosige.apps.financeiro.models import PlanoContasGrupo, PlanoContasSubgrupo
 from djangosige.apps.financeiro.forms import PlanoContasGrupoForm, PlanoContasSubgrupoFormSet
 
 
-class PlanoContasView(TemplateView):
+class PlanoContasView(CustomTemplateView):
     template_name = "financeiro/plano/plano.html"
     success_url = reverse_lazy('financeiro:planocontasview')
+    permission_codename = 'view_planocontasgrupo'
 
     def get_context_data(self, **kwargs):
         context = super(PlanoContasView, self).get_context_data(**kwargs)
@@ -31,49 +31,51 @@ class PlanoContasView(TemplateView):
 
     # Remover items selecionados da database
     def post(self, request, *args, **kwargs):
-        for key, value in request.POST.items():
-            if value == 'on':
-                grupo = None
-                subgrupo = False
-                tipo = None
-                try:
-                    instance = PlanoContasSubgrupo.objects.get(id=key)
-                    grupo = instance.grupo
-                    subgrupo = True
-                except PlanoContasGrupo.DoesNotExist:
-                    instance = PlanoContasGrupo.objects.get(id=key)
-                    grupo = instance
+        if self.check_user_delete_permission(request, PlanoContasGrupo):
+            for key, value in request.POST.items():
+                if value == 'on':
+                    grupo = None
+                    subgrupo = False
+                    tipo = None
+                    try:
+                        instance = PlanoContasSubgrupo.objects.get(id=key)
+                        grupo = instance.grupo
+                        subgrupo = True
+                    except PlanoContasGrupo.DoesNotExist:
+                        instance = PlanoContasGrupo.objects.get(id=key)
+                        grupo = instance
 
-                tipo = instance.tipo_grupo
-                instance.delete()
+                    tipo = instance.tipo_grupo
+                    instance.delete()
 
-                # Reordenar codigos dos subgrupos
-                if grupo and subgrupo:
-                    for i, obj in enumerate(PlanoContasSubgrupo.objects.filter(grupo=grupo), start=1):
-                        obj.codigo = str(grupo.codigo) + '.' + str(i)
-                        obj.save()
-                # Reordenar codigos dos grupos e subgrupos
-                else:
-                    id_list = []
-                    for g in PlanoContasGrupo.objects.filter(tipo_grupo=tipo):
-                        if not PlanoContasSubgrupo.objects.filter(id=g.id).count():
-                            id_list.append(g.id)
+                    # Reordenar codigos dos subgrupos
+                    if grupo and subgrupo:
+                        for i, obj in enumerate(PlanoContasSubgrupo.objects.filter(grupo=grupo), start=1):
+                            obj.codigo = str(grupo.codigo) + '.' + str(i)
+                            obj.save()
+                    # Reordenar codigos dos grupos e subgrupos
+                    else:
+                        id_list = []
+                        for g in PlanoContasGrupo.objects.filter(tipo_grupo=tipo):
+                            if not PlanoContasSubgrupo.objects.filter(id=g.id).count():
+                                id_list.append(g.id)
 
-                    for i, obj in enumerate(PlanoContasGrupo.objects.filter(pk__in=id_list), start=1):
-                        obj.codigo = str(i)
-                        obj.save()
-                        for j, subobj in enumerate(PlanoContasSubgrupo.objects.filter(grupo=obj), start=1):
-                            subobj.codigo = str(obj.codigo) + '.' + str(j)
-                            subobj.save()
+                        for i, obj in enumerate(PlanoContasGrupo.objects.filter(pk__in=id_list), start=1):
+                            obj.codigo = str(i)
+                            obj.save()
+                            for j, subobj in enumerate(PlanoContasSubgrupo.objects.filter(grupo=obj), start=1):
+                                subobj.codigo = str(obj.codigo) + '.' + str(j)
+                                subobj.save()
 
         return redirect(self.success_url)
 
 
-class AdicionarGrupoPlanoContasView(CreateView):
+class AdicionarGrupoPlanoContasView(CustomCreateView):
     form_class = PlanoContasGrupoForm
     template_name = "financeiro/plano/grupo_add.html"
     success_url = reverse_lazy('financeiro:planocontasview')
     success_message = "Grupo <b>%(descricao)s </b>adicionado com sucesso."
+    permission_codename = 'add_planocontasgrupo'
 
     def get_success_message(self, cleaned_data):
         return self.success_message % dict(cleaned_data, descricao=self.object.descricao)
@@ -114,24 +116,16 @@ class AdicionarGrupoPlanoContasView(CreateView):
 
             return self.form_valid(form)
 
-        return self.form_invalid(form, subgrupo_form)
-
-    def form_valid(self, form):
-        super(AdicionarGrupoPlanoContasView, self).form_valid(form)
-        messages.success(
-            self.request, self.get_success_message(form.cleaned_data))
-        return redirect(self.success_url)
-
-    def form_invalid(self, form, subgrupo_form):
-        return self.render_to_response(self.get_context_data(form=form, formset=subgrupo_form))
+        return self.form_invalid(form=form, subgrupo_form=subgrupo_form)
 
 
-class EditarGrupoPlanoContasView(UpdateView):
+class EditarGrupoPlanoContasView(CustomUpdateView):
     form_class = PlanoContasGrupoForm
     model = PlanoContasGrupo
     template_name = "financeiro/plano/grupo_edit.html"
     success_url = reverse_lazy('financeiro:planocontasview')
     success_message = "Grupo <b>%(descricao)s </b>editado com sucesso."
+    permission_codename = 'change_planocontasgrupo'
 
     def get_success_message(self, cleaned_data):
         return self.success_message % dict(cleaned_data, descricao=self.object.descricao)
@@ -163,7 +157,7 @@ class EditarGrupoPlanoContasView(UpdateView):
             self.object.save()
 
             subgrupo_form.instance = self.object
-            objs = subgrupo_form.save()
+            subgrupo_form.save()
 
             for i, obj in enumerate(PlanoContasSubgrupo.objects.filter(grupo=self.object), start=1):
                 obj.codigo = str(self.object.codigo) + '.' + str(i)
@@ -172,13 +166,4 @@ class EditarGrupoPlanoContasView(UpdateView):
 
             return self.form_valid(form)
 
-        return self.form_invalid(form, subgrupo_form)
-
-    def form_valid(self, form):
-        super(EditarGrupoPlanoContasView, self).form_valid(form)
-        messages.success(
-            self.request, self.get_success_message(form.cleaned_data))
-        return redirect(self.success_url)
-
-    def form_invalid(self, form, subgrupo_form):
-        return self.render_to_response(self.get_context_data(form=form, formset=subgrupo_form))
+        return self.form_invalid(form=form, subgrupo_form=subgrupo_form)
